Exceptions refactoring
