+++ src/main/java/svydovets/core/exception/AutowireBeanException.java
@@ -1,9 +1,6 @@
 package svydovets.core.exception;
 
 public class AutowireBeanException extends RuntimeException{
-    public AutowireBeanException(String message) {
-        super(message);
-    }
 
     public AutowireBeanException(String message, Throwable cause) {
         super(message, cause);
