+++ src/main/java/svydovets/core/exception/BeanDefinitionCreateException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class BeanDefinitionCreateException extends RuntimeException {
 
     public BeanDefinitionCreateException(String message) {
         super(message);
-    }
-
-    public BeanDefinitionCreateException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
