+++ src/main/java/svydovets/core/exception/FieldValueIllegalAccessException.java
@@ -1,10 +1,6 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class FieldValueIllegalAccessException extends RuntimeException {
-
-    public FieldValueIllegalAccessException(String message) {
-        super(message);
-    }
 
     public FieldValueIllegalAccessException(String message, Throwable cause) {
         super(message, cause);
