+++ src/main/java/svydovets/core/exception/InjectCollectionFieldException.java
@@ -1,10 +1,6 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class InjectCollectionFieldException extends RuntimeException {
-
-    public InjectCollectionFieldException(String message) {
-        super(message);
-    }
 
     public InjectCollectionFieldException(String message, Throwable cause) {
         super(message, cause);
