+++ src/main/java/svydovets/core/exception/NoSuchBeanDefinitionException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class NoSuchBeanDefinitionException extends RuntimeException {
 
     public NoSuchBeanDefinitionException(String message) {
         super(message);
-    }
-
-    public NoSuchBeanDefinitionException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
