+++ src/main/java/svydovets/core/exception/NoUniqueBeanDefinitionException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class NoUniqueBeanDefinitionException extends RuntimeException {
 
     public NoUniqueBeanDefinitionException(String message) {
         super(message);
-    }
-
-    public NoUniqueBeanDefinitionException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
