+++ src/main/java/svydovets/core/exception/NoUniqueBeanException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class NoUniqueBeanException extends RuntimeException {
 
     public NoUniqueBeanException(String message) {
         super(message);
-    }
-
-    public NoUniqueBeanException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
