+++ src/main/java/svydovets/core/exception/NoUniquePostConstructException.java
@@ -1,9 +1,8 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class NoUniquePostConstructException extends RuntimeException {
 
   public NoUniquePostConstructException(String message) {
     super(message);
   }
-
 }
