+++ src/main/java/svydovets/core/exception/UnresolvedCircularDependencyException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class UnresolvedCircularDependencyException extends RuntimeException {
 
     public UnresolvedCircularDependencyException(String message) {
         super(message);
-    }
-
-    public UnresolvedCircularDependencyException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
