+++ src/main/java/svydovets/exception/ExtractRequestBodyException.java
@@ -1,5 +1,6 @@
 package svydovets.exception;
 
+//Never used
 public class ExtractRequestBodyException extends RuntimeException {
 
     public ExtractRequestBodyException(String message) {
