+++ src/main/java/svydovets/util/exception/NoDefaultConstructorException.java
@@ -1,10 +1,6 @@
-package svydovets.exception;
+package svydovets.util.exception;
 
 public class NoDefaultConstructorException extends RuntimeException {
-
-    public NoDefaultConstructorException(String message) {
-        super(message);
-    }
 
     public NoDefaultConstructorException(String message, Throwable cause) {
         super(message, cause);
