+++ src/main/java/svydovets/web/exception/NoSuchPathVariableException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.web.exception;
 
 public class NoSuchPathVariableException extends RuntimeException {
 
     public NoSuchPathVariableException(String message) {
         super(message);
-    }
-
-    public NoSuchPathVariableException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
