+++ src/main/java/svydovets/web/exception/NoUniquePatternFoundException.java
@@ -1,12 +1,8 @@
-package svydovets.exception;
+package svydovets.web.exception;
 
 public class NoUniquePatternFoundException extends RuntimeException {
 
     public NoUniquePatternFoundException(String message) {
         super(message);
-    }
-
-    public NoUniquePatternFoundException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
