+++ src/main/java/svydovets/web/exception/ParseRequestBodyException.java
@@ -1,10 +1,6 @@
-package svydovets.exception;
+package svydovets.web.exception;
 
 public class ParseRequestBodyException extends RuntimeException {
-
-    public ParseRequestBodyException(String message) {
-        super(message);
-    }
 
     public ParseRequestBodyException(String message, Throwable cause) {
         super(message, cause);
