+++ src/main/java/svydovets/web/exception/RequestProcessingException.java
@@ -1,9 +1,6 @@
-package svydovets.exception;
+package svydovets.web.exception;
 
 public class RequestProcessingException extends RuntimeException {
-    public RequestProcessingException(String message) {
-        super(message);
-    }
 
     public RequestProcessingException(String message, Throwable cause) {
         super(message, cause);
