+++ src/main/java/svydovets/web/exception/UnsupportedTypeException.java
@@ -1,11 +1,7 @@
-package svydovets.exception;
+package svydovets.web.exception;
 
 public class UnsupportedTypeException extends RuntimeException {
     public UnsupportedTypeException(String message) {
         super(message);
-    }
-
-    public UnsupportedTypeException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
