+++ src/test/java/com/bobocode/svydovets/web/path/PathFinderTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
-import svydovets.exception.NoMatchingPatternFoundException;
-import svydovets.exception.NoUniquePatternFoundException;
+import svydovets.web.exception.NoMatchingPatternFoundException;
+import svydovets.web.exception.NoUniquePatternFoundException;
 import svydovets.web.path.PathFinder;
 import svydovets.web.path.PathFinderImpl;
 import svydovets.web.path.RequestPathParser;
